refactor: extract CombatFeedbackView + CardHandView from CombatUIController
CombatUIController was a 1473-line monolith mixing UI construction, hand
management, feedback popups, and HUD sync. This splits it into focused
components following the decomposition plan in COMBAT_ARCHITECTURE.md:

Phase 1 — CombatFeedbackView (357 lines):
  - WEAK/RESIST/MOMENTUM popup text
  - Enemy damage shake + hit animation
  - Victory/defeat animated text
  - Panel flash (energy/world) on invalid actions
  - Hand limit toast
  - Subscribes to TurnManager events independently

Phase 2 — CardHandView (400 lines):
  - Card button creation, sync, and rebuild
  - Click handler (TryPrepareCardPlay → ResolvePreparedCardPlay)
  - Attack animation coordination
  - Adaptive hand layout (scales cards when hand is large)
  - Staggered fade-in animation on draw

CombatUIController (now 980 lines) retains BuildUI orchestration, HUD
info sync, world visuals, and avatar/background rendering. It creates
both extracted views via InitializeExtractedViews() after BuildUI().

Integration pattern: both views are added as components on the same
GameObject and receive UI references via Initialize(). Neither view
mutates gameplay state — presentation only.

Also fixes BossAct1.asset: boss_recover (Regenerate) was using
effectType DrawCards (no-op for enemies); changed to Block 10.
